requirements.txt
aiogram==2.24
python-dotenv==1.0.0
spacy==3.6.0
langdetect==1.0.9
